Add -Wno-unknown-warning-option gcc PG_CPPFLAGS
Commit 01e249b0a71 made -Wno-maybe-uninitialized in PG_CPPFLAGS depend
on whether gcc is used for compilation. With PG 11 and JIT enabled, that
is however not enough, because even with CC=gcc, the same PG_CPPFLAGS
are used with clang for producing the .bc bitcode files, and clang 6.0
chokes on -Wno-maybe-uninitialized.

Fix by adding -Wno-unknown-warning-option, which will make clang ignore
unknown warning options. gcc doesn't know that option, but ignores
unknown -Wno-* options by default, so both compilers are happy.