Fix hang if many jobs are waiting to run
When more than "cron.max_running_jobs" jobs were waiting to run,
it could happen that all available slots in "pollFDs" were taken up
by waiting jobs with pendingRunCount > 0 that did nothing
(polled for file descriptor -1).

These then "crowded out" other jobs that were running, so that their
sockets never got pulled and we never detected that they are done.

In that case, pg_cron would repeat this "no-op" pull indefinitely,
causing a hang.

The typical symptom is a number of open database connections that
are idle, because their job is done.

In passing, refactor the code to make it more readable.

This should fix issue #63.