Fix hung state when cron.use_background_workers = true and the message queue is full.
With cron.use_background_workers = true, a running job can be
in a hung state waiting for the MessageQueuePutMessage wait event.
This state is reached if the message queue is FULL and the job is still running.
Indeed, currently, we are not reading the message queue if the job is still running.

This fix is reading the queue periodically and display intermediate messages (if any) in the audit table.