Fix jobname field handling
After job name was converted to a text field in 6fc9ca9 there is a
problem in updating the extension from the versions where the field is of
type name. Namely: when postgres is started but the extenision is still not
updated, thus, the column type is not yet altered, pg_cron launcher gets a
Segmentation Fault during the attempt to convert a heap tuple from
cron.job table to a CronJob struct.