Test latest version instead of hardcoding 1.4
The 1.4 version of the SQL scripts are no longer compatible with the C
code since cron_unschedule_named expects "text" as input now.

Instead of hardcoding '1.4', simply upgrade to the most recent version.
Also drop the test for extversion after the upgrade since this value is
prone to be missed on future releases.

Problem was visible on big-endian architectures.

https://github.com/citusdata/pg_cron/issues/267