Using xact_callback to catch table change event
Register the xact_callback function pgcron_xact_callback
to record if the data in the data has changed.

when transaction commits or aborts, reset the JobTableChanged
when transaction commits, if the JobTableChanged is true,
set CronJobCacheValid to false.

One drawback of this method is that in two_phase transaction,
if the transaction rollback, it will still be treat as
commited as there is no such transaction event for it
in postgres now.