PG_MODULE_MAGIC_EXT: use version from CHANGELOG.md
This causes make to take the latest version from the top of the
CHANGELOG.md file and use it for PG_MODULE_MAGIC_EXT.

This has the potential to reduce the complexity of doing new releases,
by going from three places that the version needs to be synced (git tag,
changelog, pg_cron.c) to two places (git tag and changelog), which
already was the status quo.

It does rely on the assumption that the git tag and the changelog are
kept in sync for every release, which may or may not hold forever.

The alternative would be to get the pg_cron version from git tag, but
that'll be a problem when one is building from a .tar.gz file, which may
be done by distributions nowadays.